console/requirements.txt
fastapi[all]~=0.88.0
uvicorn[standard]~=0.20.0
pydantic~=1.10.2
kafka-python~=2.0.2